refactor: rename package org.floci.core to io.floci.core
