+++ src/main/java/io/floci/core/infra/HealthCheckHandler.java
@@ -1,0 +1,8 @@
+package io.floci.core.infra;
+
+/**
+ * Interface for probing emulator runtime health status.
+ */
+public interface HealthCheckHandler {
+    HealthStatus getHealthStatus();
+}
