+++ src/main/java/io/floci/core/infra/JsonUtils.java
@@ -1,0 +1,27 @@
+package io.floci.core.infra;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+
+/**
+ * Factory and configuration helper for Jackson ObjectMapper instances across emulator services.
+ */
+public class JsonUtils {
+    private static final ObjectMapper DEFAULT_MAPPER = createMapper();
+
+    public static ObjectMapper createMapper() {
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.registerModule(new JavaTimeModule());
+        mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        mapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
+        mapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
+        return mapper;
+    }
+
+    public static ObjectMapper getDefaultMapper() {
+        return DEFAULT_MAPPER;
+    }
+}
