+++ src/main/java/io/floci/core/infra/Resettable.java
@@ -1,0 +1,12 @@
+package io.floci.core.infra;
+
+/**
+ * Interface implemented by emulator components that support state reset.
+ */
+@FunctionalInterface
+public interface Resettable {
+    /**
+     * Resets the internal in-memory state of the component.
+     */
+    void reset();
+}
