+++ src/main/java/io/floci/core/resource/compute/ComputeFunction.java
@@ -1,0 +1,12 @@
+package io.floci.core.resource.compute;
+
+/**
+ * Interface contract for serverless function invocation.
+ */
+@FunctionalInterface
+public interface ComputeFunction {
+    /**
+     * Executes a serverless function payload and returns the result.
+     */
+    ExecutionResult invoke(ExecutionPayload payload);
+}
