--- src/main/java/org/floci/core/infra/HealthCheckHandler.java
@@ -1,8 +1,0 @@
-package org.floci.core.infra;
-
-/**
- * Interface for probing emulator runtime health status.
- */
-public interface HealthCheckHandler {
-    HealthStatus getHealthStatus();
-}
