--- src/main/java/org/floci/core/infra/Resettable.java
@@ -1,12 +1,0 @@
-package org.floci.core.infra;
-
-/**
- * Interface implemented by emulator components that support state reset.
- */
-@FunctionalInterface
-public interface Resettable {
-    /**
-     * Resets the internal in-memory state of the component.
-     */
-    void reset();
-}
